+++ lightly_studio/src/lightly_studio/core/image/add_images.py
@@ -8,19 +8,16 @@ from collections import defaultdict
 from collections.abc import Mapping
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Iterable
+from typing import Iterable, Literal, cast
 from uuid import UUID
 
 import fsspec
 import PIL
-from labelformat.model.binary_mask_segmentation import BinaryMaskSegmentation
-from labelformat.model.bounding_box import BoundingBoxFormat
 from labelformat.model.image import Image
 from labelformat.model.instance_segmentation import (
     ImageInstanceSegmentation,
     InstanceSegmentationInput,
 )
-from labelformat.model.multipolygon import MultiPolygon
 from labelformat.model.object_detection import (
     ImageObjectDetection,
     ObjectDetectionInput,
@@ -28,17 +25,16 @@ from labelformat.model.object_detection import (
 from sqlmodel import Session
 from tqdm import tqdm
 
-from lightly_studio.core.image_sample import ImageSample
+from lightly_studio.core import labelformat_helpers
+from lightly_studio.core.image.image_sample import ImageSample
 from lightly_studio.core.loading_log import LoadingLoggingContext, log_loading_results
 from lightly_studio.models.annotation.annotation_base import (
     AnnotationCreate,
     AnnotationType,
 )
-from lightly_studio.models.annotation_label import AnnotationLabelCreate
 from lightly_studio.models.caption import CaptionCreate
 from lightly_studio.models.image import ImageCreate
 from lightly_studio.resolvers import (
-    annotation_label_resolver,
     annotation_resolver,
     caption_resolver,
     image_resolver,
@@ -75,6 +71,7 @@ def load_into_dataset_from_paths(
     session: Session,
     dataset_id: UUID,
     image_paths: Iterable[str],
+    show_progress: bool = True,
 ) -> list[UUID]:
     """Load images from file paths into the dataset.
 
@@ -82,6 +79,7 @@ def load_into_dataset_from_paths(
         session: The database session.
         dataset_id: The ID of the dataset to load images into.
         image_paths: An iterable of file paths to the images to load.
+        show_progress: Whether to display a progress bar and final summary of loading results.
 
     Returns:
         A list of UUIDs of the created samples.
@@ -100,6 +98,7 @@ def load_into_dataset_from_paths(
         image_paths,
         desc="Processing images",
         unit=" images",
+        disable=not show_progress,
     ):
         try:
             with fsspec.open(image_path, "rb") as file:
@@ -134,7 +133,12 @@ def load_into_dataset_from_paths(
         created_sample_ids.extend(created_path_to_id.values())
         logging_context.update_example_paths(paths_not_inserted)
 
-    log_loading_results(session=session, dataset_id=dataset_id, logging_context=logging_context)
+    log_loading_results(
+        session=session,
+        dataset_id=dataset_id,
+        logging_context=logging_context,
+        print_summary=show_progress,
+    )
     return created_sample_ids
 
 
@@ -238,7 +242,9 @@ def load_into_dataset_from_labelformat(
             label_map=label_map,
         )
 
-    log_loading_results(session=session, dataset_id=dataset_id, logging_context=logging_context)
+    log_loading_results(
+        session=session, dataset_id=dataset_id, logging_context=logging_context, print_summary=True
+    )
     return created_sample_ids
 
 
@@ -335,7 +341,9 @@ def load_into_dataset_from_coco_captions(
             path_to_captions=path_to_captions,
         )
 
-    log_loading_results(session=session, dataset_id=dataset_id, logging_context=logging_context)
+    log_loading_results(
+        session=session, dataset_id=dataset_id, logging_context=logging_context, print_summary=True
+    )
     return created_sample_ids
 
 
@@ -429,26 +437,11 @@ def _create_label_map(
         dataset_id: The ID of the root collection the labels belong to.
         input_labels: The labelformat input containing the categories.
     """
-    label_map = {}
-    for category in tqdm(
-        input_labels.get_categories(),
-        desc="Processing categories",
-        unit=" categories",
-    ):
-        # Use label if already exists
-        label = annotation_label_resolver.get_by_label_name(
-            session=session, dataset_id=dataset_id, label_name=category.name
-        )
-        if label is None:
-            # Create new label
-            label_create = AnnotationLabelCreate(
-                dataset_id=dataset_id,
-                annotation_label_name=category.name,
-            )
-            label = annotation_label_resolver.create(session=session, label=label_create)
-
-        label_map[category.id] = label.annotation_label_id
-    return label_map
+    return labelformat_helpers.create_label_map(
+        session=session,
+        dataset_id=dataset_id,
+        input_labels=input_labels,
+    )
 
 
 def _process_object_detection_annotations(
@@ -464,18 +457,11 @@ def _process_object_detection_annotations(
 
     new_annotations = []
     for obj in anno_data.data.objects:
-        box = obj.box.to_format(BoundingBoxFormat.XYWH)
-        x, y, width, height = box
-
         new_annotations.append(
-            AnnotationCreate(
+            labelformat_helpers.get_object_detection_annotation_create(
                 parent_sample_id=context.sample_id,
                 annotation_label_id=context.label_map[obj.category.id],
-                annotation_type=AnnotationType.OBJECT_DETECTION,
-                x=int(x),
-                y=int(y),
-                width=int(width),
-                height=int(height),
+                box=obj.box,
                 confidence=obj.confidence,
             )
         )
@@ -493,29 +479,18 @@ def _process_segmentation_annotations(
     ):
         raise ValueError("Invalid annotation data for segmentation processing.")
 
+    annotation_type = cast(
+        Literal[AnnotationType.INSTANCE_SEGMENTATION, AnnotationType.SEMANTIC_SEGMENTATION],
+        anno_data.annotation_type,
+    )
     new_annotations = []
     for obj in anno_data.data.objects:
-        segmentation_rle: None | list[int] = None
-        if isinstance(obj.segmentation, MultiPolygon):
-            box = obj.segmentation.bounding_box().to_format(BoundingBoxFormat.XYWH)
-        elif isinstance(obj.segmentation, BinaryMaskSegmentation):
-            box = obj.segmentation.bounding_box.to_format(BoundingBoxFormat.XYWH)
-            segmentation_rle = obj.segmentation._rle_row_wise  # noqa: SLF001
-        else:
-            raise ValueError(f"Unsupported segmentation type: {type(obj.segmentation)}")
-
-        x, y, width, height = box
-
         new_annotations.append(
-            AnnotationCreate(
+            labelformat_helpers.get_segmentation_annotation_create(
                 parent_sample_id=context.sample_id,
                 annotation_label_id=context.label_map[obj.category.id],
-                annotation_type=anno_data.annotation_type,
-                x=int(x),
-                y=int(y),
-                width=int(width),
-                height=int(height),
-                segmentation_mask=segmentation_rle,
+                segmentation=obj.segmentation,
+                annotation_type=annotation_type,
             )
         )
     return new_annotations